chore: separate filtering section and context provider
