+++ src/components/dimension-modal/conditions-modal-content/filtering-section.tsx
@@ -3,106 +3,29 @@ import { useFilterRadioMode } from '@components/dimension-modal/show-all-filter-
 import { valueTypeDisplayNames } from '@dhis2/analytics'
 import i18n from '@dhis2/d2-i18n'
 import { Button } from '@dhis2/ui'
-import { useAppDispatch, useAppSelector } from '@hooks'
-import {
-    parseConditionsArrayToString,
-    parseConditionsStringToArray,
-} from '@modules/conditions'
-import {
-    getVisUiConfigConditionsByDimension,
-    setVisUiConfigConditionsByDimension,
-} from '@store/vis-ui-config-slice'
-import type { DimensionMetadataItem, ValueType } from '@types'
-import { type FC, useCallback, useMemo, useState } from 'react'
+import { type FC, useCallback } from 'react'
 import { Conditions } from './conditions'
-import {
-    ConditionsContext,
-    type ConditionsContextValue,
-} from './conditions-context'
+import { useConditions } from './conditions-provider'
 import classes from './styles/conditions-modal-content.module.css'
 
-const EMPTY_CONDITION = ''
-
-const SINGLETON_TYPES: Set<ValueType> = new Set([
-    'BOOLEAN',
-    'TRUE_ONLY',
-    'ORGANISATION_UNIT',
-])
-
-// List of supported types for conditions
-const SUPPORTED_TYPES: Set<ValueType> = new Set([
-    'NUMBER',
-    'UNIT_INTERVAL',
-    'PERCENTAGE',
-    'INTEGER',
-    'INTEGER_POSITIVE',
-    'INTEGER_NEGATIVE',
-    'INTEGER_ZERO_OR_POSITIVE',
-    'TEXT',
-    'LONG_TEXT',
-    'LETTER',
-    'PHONE_NUMBER',
-    'EMAIL',
-    'USERNAME',
-    'URL',
-    'BOOLEAN',
-    'TRUE_ONLY',
-    'DATE',
-    'TIME',
-    'DATETIME',
-    'ORGANISATION_UNIT',
-])
-
 type FilteringSectionProps = {
-    dimension: DimensionMetadataItem
     showHeading: boolean
 }
 
 export const FilteringSection: FC<FilteringSectionProps> = ({
-    dimension,
     showHeading,
 }) => {
-    const dispatch = useAppDispatch()
-
-    const conditions = useAppSelector((state) =>
-        getVisUiConfigConditionsByDimension(state, dimension?.id)
-    )
-
-    const valueType = dimension.valueType
-    const isProgramIndicator: boolean =
-        dimension.dimensionType === 'PROGRAM_INDICATOR'
-    const isLegendSetCondition: boolean = Boolean(conditions.legendSet)
-    const isOptionSetCondition: boolean = Boolean(dimension.optionSetId)
-    const isSingleCondition: boolean = Boolean(
-        isLegendSetCondition ||
-        isOptionSetCondition ||
-        (valueType && SINGLETON_TYPES.has(valueType))
-    )
-    const isSupported: boolean = Boolean(
-        isProgramIndicator || (valueType && SUPPORTED_TYPES.has(valueType))
-    )
-
-    const [conditionsList, setConditionsList] = useState<string[]>(
-        conditions.condition?.length
-            ? parseConditionsStringToArray(conditions.condition)
-            : [EMPTY_CONDITION]
-    )
-
-    const storeConditions = useCallback(
-        (conditionsList: string[]) =>
-            dispatch(
-                setVisUiConfigConditionsByDimension({
-                    dimensionId: dimension.id,
-                    conditions: parseConditionsArrayToString(
-                        conditionsList.filter(
-                            (condition) =>
-                                condition.length && !condition.endsWith(':')
-                        )
-                    ),
-                })
-            ),
-        [dimension.id, dispatch]
-    )
+    const {
+        dimension,
+        conditions,
+        conditionsList,
+        valueType,
+        isLegendSetCondition,
+        isSingleCondition,
+        isSupported,
+        addCondition,
+        storeConditions,
+    } = useConditions()
 
     const onEnterShowAll = useCallback(
         () => storeConditions([]),
@@ -120,66 +43,6 @@ export const FilteringSection: FC<FilteringSectionProps> = ({
         onEnterFilter,
     })
 
-    const addCondition = (): void => {
-        setConditionsList((prev) => [...prev, EMPTY_CONDITION])
-    }
-
-    const removeCondition = useCallback<
-        ConditionsContextValue['removeCondition']
-    >(
-        (conditionIndex) =>
-            setConditionsList((prev) => {
-                const updatedConditionsList = prev.filter(
-                    (_, index) => index !== conditionIndex
-                )
-
-                storeConditions(updatedConditionsList)
-
-                return updatedConditionsList
-            }),
-        [storeConditions]
-    )
-
-    const setCondition = useCallback<ConditionsContextValue['setCondition']>(
-        (conditionIndex, value) =>
-            setConditionsList((prev) => {
-                const updatedConditionsList = prev.map((condition, index) =>
-                    index === conditionIndex ? value : condition
-                )
-
-                storeConditions(updatedConditionsList)
-
-                return updatedConditionsList
-            }),
-        [storeConditions]
-    )
-
-    const contextValue: ConditionsContextValue = useMemo(() => {
-        return {
-            dimension,
-            conditions,
-            conditionsList,
-            isLegendSetCondition,
-            isOptionSetCondition,
-            isProgramIndicator,
-            isSupported,
-            valueType,
-            setCondition,
-            removeCondition,
-        }
-    }, [
-        dimension,
-        conditions,
-        conditionsList,
-        isLegendSetCondition,
-        isOptionSetCondition,
-        isProgramIndicator,
-        isSupported,
-        valueType,
-        setCondition,
-        removeCondition,
-    ])
-
     const filterDisabledHelp = valueType
         ? i18n.t('{{valueType}} type dimensions cannot be filtered.', {
               valueType: valueTypeDisplayNames[valueType],
@@ -191,42 +54,42 @@ export const FilteringSection: FC<FilteringSectionProps> = ({
         ? i18n.t('Show all groups')
         : undefined
 
+    if (!isSupported) {
+        return (
+            <ShowAllFilterRadio
+                mode="SHOW_ALL"
+                onModeChange={() => {
+                    /* unfilterable dimensions are always "Show all" */
+                }}
+                heading={heading}
+                dataTest={`conditions-${dimension.id}-filter-radio`}
+                filterDisabled
+                filterDisabledHelp={filterDisabledHelp}
+            />
+        )
+    }
+
     return (
-        <ConditionsContext.Provider value={contextValue}>
-            {isSupported ? (
-                <ShowAllFilterRadio
-                    mode={mode}
-                    onModeChange={onModeChange}
-                    heading={heading}
-                    showAllLabel={showAllLabel}
-                    dataTest={`conditions-${dimension.id}-filter-radio`}
-                >
-                    <div className={classes.mainSection}>
-                        <Conditions />
-                        {!isSingleCondition && (
-                            <Button
-                                type="button"
-                                small
-                                onClick={addCondition}
-                                dataTest="button-add-condition"
-                            >
-                                {i18n.t('Add filter')}
-                            </Button>
-                        )}
-                    </div>
-                </ShowAllFilterRadio>
-            ) : (
-                <ShowAllFilterRadio
-                    mode="SHOW_ALL"
-                    onModeChange={() => {
-                        /* unfilterable dimensions are always "Show all" */
-                    }}
-                    heading={heading}
-                    dataTest={`conditions-${dimension.id}-filter-radio`}
-                    filterDisabled
-                    filterDisabledHelp={filterDisabledHelp}
-                />
-            )}
-        </ConditionsContext.Provider>
+        <ShowAllFilterRadio
+            mode={mode}
+            onModeChange={onModeChange}
+            heading={heading}
+            showAllLabel={showAllLabel}
+            dataTest={`conditions-${dimension.id}-filter-radio`}
+        >
+            <div className={classes.mainSection}>
+                <Conditions />
+                {!isSingleCondition && (
+                    <Button
+                        type="button"
+                        small
+                        onClick={addCondition}
+                        dataTest="button-add-condition"
+                    >
+                        {i18n.t('Add filter')}
+                    </Button>
+                )}
+            </div>
+        </ShowAllFilterRadio>
     )
 }
